rename processRepo.ts to updateProject.ts
