+++ packages/graphql_server/src/supabaseUtils.ts
@@ -1,14 +1,14 @@
 import supabase from './supabase'
 import { getOrganizationInfo, getUserInfo } from './api/githubApi'
+/*
+Types:
+*/
 import {
   OrganizationInsertion,
   OrganizationUpdate,
   PersonInsertion,
   ProjectUpdate
 } from '../types/dataAggregation'
-/*
-Types:
-*/
 import { GitHubOrganization, GitHubInfo, GitHubUser } from '../types/githubApi'
 import { LinkedInCompanyProfile } from '../types/linkedInScraping'
 import { StarRecord } from '../types/starHistory'
